Enrichments reorganization
* Add `normalizatinos` to `enrichments` package
* Move string munging "enrichments" into `normalizations` package
	* StringUtils
	* Filter
	* StandardStringEnrichments
	* DeDuplication
