+++ src/test/scala/dpla/ingestion3/enrichments/FilterRegexTest.scala
@@ -1,6 +1,6 @@
 package dpla.ingestion3.enrichments
 
-import dpla.ingestion3.enrichments.FilterRegex._
+import dpla.ingestion3.enrichments.normalizations.FilterRegex._
 import org.scalatest._
 
 import scala.util.matching
