+++ src/test/scala/dpla/ingestion3/enrichments/StringUtilsTest.scala
@@ -1,6 +1,7 @@
 package dpla.ingestion3.enrichments
 
-import dpla.ingestion3.enrichments.StringUtils._
+import dpla.ingestion3.enrichments.normalizations.StringUtils._
+import dpla.ingestion3.enrichments.normalizations.FilterList
 import org.scalatest.{BeforeAndAfter, FlatSpec}
 
 class StringUtilsTest extends FlatSpec with BeforeAndAfter {
